SDK update
-Assembly Reference integration for Windows Handler
