+++ MagicUnityTemplate/Assets/MagicUnity/link.magic.unity.sdk/Scripts/Windows_Handler.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c9da25a2095d8e489b26f400b7dd74d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
