+++ MagicUnityTemplate/Assets/MagicUnity/link.magic.unity.sdk/Scripts/Windows_Handler/Windows_Handler.cs.meta
