+++ MagicUnityTemplate/Assets/MagicUnity/link.magic.unity.sdk/Scripts/link.magic.unity.sdk.asmdef
@@ -1,3 +1,16 @@
 {
-  "name": "link.magic.unity.sdk"
+    "name": "link.magic.unity.sdk",
+    "rootNamespace": "",
+    "references": [
+        "GUID:9e287cc7b3c8a0245a7d36254472d3c8"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
